+++ otterdog/models/ruleset.py
@@ -39,6 +39,79 @@ RS = TypeVar("RS", bound="Ruleset")
 
 
 @dataclasses.dataclass
+class PullRequestSettings(EmbeddedModelObject):
+    required_approving_review_count: int
+    dismisses_stale_reviews: bool
+    requires_code_owner_review: bool
+    requires_last_push_approval: bool
+    requires_review_thread_resolution: bool
+
+    def validate(self, context: ValidationContext, parent_object: Any) -> None:
+        for key in ["required_approving_review_count"]:
+            value = self.__getattribute__(key)
+            if is_unset(value):
+                context.add_failure(
+                    FailureType.ERROR,
+                    f"{parent_object.get_model_header(parent_object)} has not set required parameter "
+                    f"'required_pull_request.{key}'.",
+                )
+
+        if is_set_and_valid(self.required_approving_review_count):
+            if self.required_approving_review_count < 0 or self.required_approving_review_count > 10:
+                context.add_failure(
+                    FailureType.ERROR,
+                    f"{parent_object.get_model_header(parent_object)} has "
+                    f"'required_pull_request.required_approving_review_count' of value "
+                    f"'{self.required_approving_review_count}' while only integers in the range [0, 10] are allowed.",
+                )
+
+    def get_jsonnet_template_function(self, jsonnet_config: JsonnetConfig, extend: bool) -> str | None:
+        return f"orgs.{jsonnet_config.create_pull_request}"
+
+    @classmethod
+    def get_mapping_from_provider(cls, org_id: str, data: dict[str, Any]) -> dict[str, Any]:
+        mapping = super().get_mapping_from_provider(org_id, data)
+
+        mapping.update(
+            {
+                "required_approving_review_count": OptionalS("required_approving_review_count", default=UNSET),
+                "dismisses_stale_reviews": OptionalS("dismiss_stale_reviews_on_push", default=UNSET),
+                "requires_code_owner_review": OptionalS("require_code_owner_review", default=UNSET),
+                "requires_last_push_approval": OptionalS("require_last_push_approval", default=UNSET),
+                "requires_review_thread_resolution": OptionalS("required_review_thread_resolution", default=UNSET),
+            }
+        )
+
+        return mapping
+
+    @classmethod
+    async def get_mapping_to_provider(
+        cls, org_id: str, data: dict[str, Any], provider: GitHubProvider
+    ) -> dict[str, Any]:
+        mapping = super().get_mapping_from_provider(org_id, data)
+
+        mapping.update(
+            {
+                "dismiss_stale_reviews_on_push": S("dismisses_stale_reviews"),
+                "require_code_owner_review": S("requires_code_owner_review"),
+                "require_last_push_approval": S("requires_last_push_approval"),
+                "required_review_thread_resolution": S("requires_review_thread_resolution"),
+            }
+        )
+
+        for key in [
+            "dismisses_stale_reviews",
+            "requires_code_owner_review",
+            "requires_last_push_approval",
+            "requires_review_thread_resolution",
+        ]:
+            if key in mapping:
+                mapping.pop(key)
+
+        return mapping
+
+
+@dataclasses.dataclass
 class MergeQueueSettings(EmbeddedModelObject):
     merge_method: str
     build_concurrency: int
@@ -161,15 +234,6 @@ class Ruleset(ModelObject, abc.ABC):
     requires_commit_signatures: bool
     requires_linear_history: bool
 
-    requires_pull_request: bool
-    # the following settings are only taken into account
-    # when requires_pull_request is True
-    required_approving_review_count: int | None
-    dismisses_stale_reviews: bool
-    requires_code_owner_review: bool
-    requires_last_push_approval: bool
-    requires_review_thread_resolution: bool
-
     requires_status_checks: bool
     requires_strict_status_checks: bool
     required_status_checks: list[str]
@@ -177,6 +241,7 @@ class Ruleset(ModelObject, abc.ABC):
     requires_deployments: bool
     required_deployment_environments: list[str]
 
+    required_pull_request: PullRequestSettings | None = dataclasses.field(metadata={"embedded_model": True})
     required_merge_queue: MergeQueueSettings | None = dataclasses.field(metadata={"embedded_model": True})
 
     _roles: ClassVar[dict[str, str]] = {"5": "RepositoryAdmin", "4": "Write", "2": "Maintain", "1": "OrganizationAdmin"}
@@ -234,40 +299,6 @@ class Ruleset(ModelObject, abc.ABC):
                         f"'{ref}', only values ('refs/heads/*', '~DEFAULT_BRANCH', '~ALL') are allowed",
                     )
 
-        if self.requires_pull_request is False:
-            if is_set_and_valid(self.required_approving_review_count):
-                context.add_failure(
-                    FailureType.WARNING,
-                    f"{self.get_model_header(parent_object)} has"
-                    f" 'requires_pull_request' disabled but 'required_approving_review_count' "
-                    f"is set to '{self.required_approving_review_count}', setting will be ignored.",
-                )
-
-            for key in {
-                "dismisses_stale_reviews",
-                "requires_code_owner_review",
-                "requires_last_push_approval",
-                "requires_review_thread_resolution",
-            }:
-                if self.__getattribute__(key) is True:
-                    context.add_failure(
-                        FailureType.INFO,
-                        f"{self.get_model_header(parent_object)} has"
-                        f" 'requires_pull_request' disabled but '{key}' "
-                        f"is enabled, setting will be ignored.",
-                    )
-
-        # required_approving_review_count must be defined when requires_pull_request is enabled
-        required_approving_review_count = self.required_approving_review_count
-        if self.requires_pull_request is True and not is_unset(required_approving_review_count):
-            if required_approving_review_count is None or required_approving_review_count < 0:
-                context.add_failure(
-                    FailureType.ERROR,
-                    f"{self.get_model_header(parent_object)} has"
-                    f" 'requires_pull_request' enabled but 'required_approving_review_count' "
-                    f"is not set (must be set to a non negative number).",
-                )
-
         # if 'requires_status_checks' is disabled, issue a warning if required_status_checks is non-empty.
         required_status_checks = self.required_status_checks
         if (
@@ -311,21 +342,13 @@ class Ruleset(ModelObject, abc.ABC):
                         f"'{env_name}' which is not defined in the repository itself.",
                     )
 
+        if is_set_and_present(self.required_pull_request):
+            self.required_pull_request.validate(context, parent_object)
+
         if is_set_and_present(self.required_merge_queue):
             self.required_merge_queue.validate(context, parent_object)
 
     def include_field_for_diff_computation(self, field: dataclasses.Field) -> bool:
-        # disable diff computation for dependent fields of requires_pull_request,
-        if self.requires_pull_request is False:
-            if field.name in [
-                "required_approving_review_count",
-                "dismisses_stale_reviews",
-                "requires_code_owner_review",
-                "requires_last_push_approval",
-                "requires_review_thread_resolution",
-            ]:
-                return False
-
         if self.requires_status_checks is False:
             if field.name in [
                 "required_status_checks",
@@ -348,11 +371,16 @@ class Ruleset(ModelObject, abc.ABC):
 
         mapping.update(
             {
+                "required_pull_request": If(
+                    OptionalS("required_pull_request", default=None) == K(None),
+                    K(None),
+                    S("required_pull_request") >> F(lambda x: PullRequestSettings.from_model_data(x)),
+                ),
                 "required_merge_queue": If(
                     OptionalS("required_merge_queue", default=None) == K(None),
                     K(None),
                     S("required_merge_queue") >> F(lambda x: MergeQueueSettings.from_model_data(x)),
-                )
+                ),
             }
         )
 
@@ -421,23 +449,12 @@ class Ruleset(ModelObject, abc.ABC):
         else:
             mapping["bypass_actors"] = K([])
 
-        # requires pull request
+        # required pull request
         if any((found := rule) for rule in rules if rule["type"] == "pull_request"):
-            mapping["requires_pull_request"] = K(True)
             parameters = found.get("parameters", {})
-
-            mapping["required_approving_review_count"] = K(parameters.get("required_approving_review_count", None))
-            mapping["dismisses_stale_reviews"] = K(parameters.get("dismiss_stale_reviews_on_push", UNSET))
-            mapping["requires_code_owner_review"] = K(parameters.get("require_code_owner_review", UNSET))
-            mapping["requires_last_push_approval"] = K(parameters.get("require_last_push_approval", UNSET))
-            mapping["requires_review_thread_resolution"] = K(parameters.get("required_review_thread_resolution", UNSET))
+            mapping["required_pull_request"] = K(PullRequestSettings.from_provider_data(org_id, parameters))
         else:
-            mapping["requires_pull_request"] = K(False)
-            mapping["required_approving_review_count"] = K(None)
-            mapping["dismisses_stale_reviews"] = K(UNSET)
-            mapping["requires_code_owner_review"] = K(UNSET)
-            mapping["requires_last_push_approval"] = K(UNSET)
-            mapping["requires_review_thread_resolution"] = K(UNSET)
+            mapping["required_pull_request"] = K(None)
 
         # required status checks
         if any((found := rule) for rule in rules if rule["type"] == "required_status_checks"):
@@ -565,34 +582,20 @@ class Ruleset(ModelObject, abc.ABC):
             if prop_key in data:
                 params[param_key] = S(prop_key)
 
-        # requires pull request
-        if "requires_pull_request" in data:
-            value = data["requires_pull_request"]
-
-            if value is True:
-                rule = {"type": K("pull_request")}
-                pull_parameters: dict[str, Any] = {}
-
-                add_parameter("required_approving_review_count", "required_approving_review_count", pull_parameters)
-                add_parameter("dismisses_stale_reviews", "dismiss_stale_reviews_on_push", pull_parameters)
-                add_parameter("requires_code_owner_review", "require_code_owner_review", pull_parameters)
-                add_parameter("requires_last_push_approval", "require_last_push_approval", pull_parameters)
-                add_parameter("requires_review_thread_resolution", "required_review_thread_resolution", pull_parameters)
-
-                rule["parameters"] = pull_parameters
+        # required merge queue
+        if "required_pull_request" in data:
+            mapping.pop("required_pull_request")
+            parameters = await PullRequestSettings.dict_to_provider_data(
+                org_id,
+                data["required_pull_request"],
+                provider,
+            )
+            if parameters and len(parameters) > 0:
+                rule = {
+                    "type": K("pull_request"),
+                    "parameters": K(parameters),
+                }
                 rules.append(rule)
-
-        pop_mapping(
-            [
-                "requires_pull_request",
-                "required_approving_review_count",
-                "dismisses_stale_reviews",
-                "requires_code_owner_review",
-                "requires_last_push_approval",
-                "requires_last_push_approval",
-                "requires_review_thread_resolution",
-            ]
-        )
 
         # required status checks
         if "requires_status_checks" in data:
@@ -690,16 +693,53 @@ class Ruleset(ModelObject, abc.ABC):
 
         printer.print(f"{template_function}('{self.name}')")
 
+        if "required_pull_request" in patch and patch.get("required_pull_request") is not None:
+            patch.pop("required_pull_request")
+
         if "required_merge_queue" in patch and patch.get("required_merge_queue") is not None:
             patch.pop("required_merge_queue")
 
         write_patch_object_as_json(patch, printer, close_object=False)
 
+        if is_set_and_present(self.required_pull_request):
+            default_pull_request_config = cast(Ruleset, default_object).required_pull_request
+            if default_pull_request_config is None:
+                default_pull_request_config = PullRequestSettings.from_model_data(
+                    jsonnet_config.default_pull_request_config
+                )
+                embedded_extend = False
+            else:
+                embedded_extend = True
+
+            if is_set_and_valid(default_pull_request_config):
+                printer.print(f"required_pull_request{'+' if embedded_extend else ''}:")
+                self.required_pull_request.to_jsonnet(
+                    printer,
+                    jsonnet_config,
+                    context,
+                    embedded_extend,
+                    default_pull_request_config,
+                )
+
         if is_set_and_present(self.required_merge_queue):
-            default_merge_queue_config = MergeQueueSettings.from_model_data(jsonnet_config.default_merge_queue_config)
+            default_merge_queue_config = cast(Ruleset, default_object).required_merge_queue
+            if default_merge_queue_config is None:
+                default_merge_queue_config = MergeQueueSettings.from_model_data(
+                    jsonnet_config.default_merge_queue_config
+                )
+                embedded_extend = False
+            else:
+                embedded_extend = True
+
             if is_set_and_valid(default_merge_queue_config):
-                printer.print("required_merge_queue: ")
-                self.required_merge_queue.to_jsonnet(printer, jsonnet_config, context, default_merge_queue_config)
+                printer.print(f"required_merge_queue{'+' if embedded_extend else ''}:")
+                self.required_merge_queue.to_jsonnet(
+                    printer,
+                    jsonnet_config,
+                    context,
+                    embedded_extend,
+                    default_merge_queue_config,
+                )
 
         # close the object
         printer.level_down()