+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 
 ### Changed
 
+- Converted pull request related settings of a Ruleset into an embedded model object similar to merge queue settings.
 - Use `jsonata` instead of `jq` for querying json objects.
 - Use `ghproxy` by default as transparent cache / proxy when accessing the GitHub API from the webapp. ([#274](https://github.com/eclipse-csi/otterdog/issues/274))
 - Changed parameter `--update-filter` for various operations from a python regular expression to a shell pattern format.